Move file_stream and file_writer declaration out of file.h
This extended modularity will become important, when we get to the BDD class, which the file_writer and file_streams then can include
