+++ src/coom/bdd/apply.cpp
@@ -3,6 +3,8 @@
 
 #include "apply.h"
 
+#include <coom/file_stream.h>
+#include <coom/file_writer.h>
 #include <coom/priority_queue.h>
 #include <coom/reduce.h>
