+++ src/coom/bdd/build.cpp
@@ -3,6 +3,9 @@
 
 #include "build.h"
 
+#include <coom/file_stream.h>
+#include <coom/file_writer.h>
+
 #include <coom/assert.h>
 
 namespace coom {
