+++ src/coom/reduce.cpp
@@ -5,7 +5,10 @@
 
 #include <tpie/sort.h>
 
+#include <coom/file_stream.h>
+#include <coom/file_writer.h>
 #include <coom/priority_queue.h>
+
 #include <coom/assert.h>
 
 namespace coom
